up state from Search to App
